refactor: add sections, official-biography
