+++ public/r/articles.json
@@ -9,8 +9,8 @@
   ],
   "files": [
     {
-      "path": "src/lib/demo/blocks/articles/index.tsx",
-      "content": "import React from 'react';\nimport { Card, CardHeader, CardContent } from '@kapwa/card';\nimport {\n  ArticleCard,\n  ArticleCardItem,\n  ArticleContentSection,\n  ArticleContent,\n} from '@kapwa/article';\n\ntype ArticlesSectionProps = {\n  headerTitle: string;\n  headerSubtitle?: string;\n  articleCardTitle?: string;\n  articleCards: ArticleCardItem[];\n  showArticleContent?: boolean;\n  articleContentTitle?: string;\n  articleContent?: ArticleContentSection[];\n};\n\nconst ArticlesSection: React.FC<ArticlesSectionProps> = ({\n  headerTitle,\n  headerSubtitle,\n  articleCardTitle = 'Article Card',\n  articleCards,\n  showArticleContent = true,\n  articleContentTitle = 'Article Content',\n  articleContent,\n}: ArticlesSectionProps) => {\n  return (\n    <Card className='mb-8'>\n      <CardHeader>\n        <h2 className='text-2xl font-semibold'>{headerTitle}</h2>\n        {headerSubtitle && <p className='text-gray-800'>{headerSubtitle}</p>}\n      </CardHeader>\n      <CardContent>\n        <div className='space-y-8'>\n          {/* Article Card */}\n          <div>\n            <h3 className='font-medium mb-4'>{articleCardTitle}</h3>\n            <div className='grid grid-cols-1 md:grid-cols-2 gap-6'>\n              {articleCards.map(article => (\n                <ArticleCard key={article.id} article={article} />\n              ))}\n            </div>\n          </div>\n\n          {/* Article Content */}\n          {showArticleContent && articleContent && (\n            <ArticleContent\n              articleContentTitle={articleContentTitle}\n              articleContent={articleContent}\n            />\n          )}\n        </div>\n      </CardContent>\n    </Card>\n  );\n};\n\nexport { ArticlesSection };\nexport type { ArticleCardItem, ArticleContentSection, ArticlesSectionProps };\n",
+      "path": "src/lib/demo/blocks/sections/articles/index.tsx",
+      "content": "import React from 'react';\nimport { Card, CardHeader, CardContent } from '@kapwa/card';\nimport {\n  ArticleCard,\n  ArticleCardItem,\n  ArticleContentSection,\n  ArticleContent,\n} from '@kapwa/article';\n\ntype ArticlesSectionProps = {\n  headerTitle: string;\n  headerSubtitle?: string;\n  articleCardTitle?: string;\n  articleCards: ArticleCardItem[];\n  showArticleContent?: boolean;\n  articleContentTitle?: string;\n  articleContent?: ArticleContentSection[];\n};\n\nconst ArticlesSection: React.FC<ArticlesSectionProps> = ({\n  headerTitle,\n  headerSubtitle,\n  articleCardTitle = 'Article Card',\n  articleCards,\n  showArticleContent = true,\n  articleContentTitle = 'Article Content',\n  articleContent,\n}: ArticlesSectionProps) => {\n  return (\n    <Card className='mb-2'>\n      <CardHeader>\n        <h2 className='text-2xl font-semibold'>{headerTitle}</h2>\n        {headerSubtitle && <p className='text-gray-800'>{headerSubtitle}</p>}\n      </CardHeader>\n      <CardContent>\n        <div className='space-y-8'>\n          {/* Article Card */}\n          <div>\n            <h3 className='font-medium mb-4'>{articleCardTitle}</h3>\n            <div className='grid grid-cols-1 md:grid-cols-2 gap-6'>\n              {articleCards.map(article => (\n                <ArticleCard key={article.id} article={article} />\n              ))}\n            </div>\n          </div>\n\n          {/* Article Content */}\n          {showArticleContent && articleContent && (\n            <ArticleContent\n              articleContentTitle={articleContentTitle}\n              articleContent={articleContent}\n            />\n          )}\n        </div>\n      </CardContent>\n    </Card>\n  );\n};\n\nexport { ArticlesSection };\nexport type { ArticleCardItem, ArticleContentSection, ArticlesSectionProps };\n",
       "type": "registry:block"
     }
   ],
