+++ public/r/list.json
@@ -10,7 +10,7 @@
   "files": [
     {
       "path": "src/lib/kapwa/list/index.tsx",
-      "content": "import { Button } from '@kapwa/button';\nimport { Card, CardHeader, CardContent } from '@kapwa/card';\nimport { ExternalLink } from 'lucide-react';\n\ntype ListSectionItem = {\n  id: number;\n  title: string;\n  category: string;\n};\n\ntype ListsSectionProps = {\n  title: string;\n  headerTitle: string;\n  headerSubtitle?: string;\n  externalLinkText?: string;\n  listItems: ListSectionItem[];\n};\n\nconst List = ({\n  headerTitle,\n  headerSubtitle,\n  externalLinkText,\n  listItems,\n  title,\n}: {\n  title: string;\n  headerTitle: string;\n  headerSubtitle?: string;\n  externalLinkText?: string;\n  listItems: ListSectionItem[];\n}) => {\n  return (\n    <Card className='mb-8'>\n      <CardHeader>\n        <h2 className='text-2xl font-semibold'>{headerTitle}</h2>\n        {headerSubtitle && <p className='text-gray-800'>{headerSubtitle}</p>}\n      </CardHeader>\n      <CardContent>\n        <div className='space-y-8'>\n          <div>\n            <h3 className='font-medium mb-4'>{title}</h3>\n            <div className='space-y-4'>\n              {listItems.map(item => (\n                <div\n                  key={item.id}\n                  className='p-4 bg-white border border-gray-200 rounded-lg hover:border-primary-500 transition-colors'\n                >\n                  <div className='flex items-start justify-between'>\n                    <div>\n                      <h4 className='text-lg font-medium text-gray-900'>\n                        {item.title}\n                      </h4>\n                      <span className='inline-block px-2 py-1 mt-2 text-xs font-medium rounded-sm bg-gray-100 text-gray-800'>\n                        {item.category}\n                      </span>\n                    </div>\n                    {externalLinkText && (\n                      <Button variant='ghost' size='sm'>\n                        {externalLinkText}\n                        <ExternalLink className='ml-2 h-4 w-4' />\n                      </Button>\n                    )}\n                  </div>\n                </div>\n              ))}\n            </div>\n          </div>\n        </div>\n      </CardContent>\n    </Card>\n  );\n};\n\nexport { List };\nexport type { ListsSectionProps, ListSectionItem };\n",
+      "content": "import { Button } from '@kapwa/button';\nimport { Card, CardHeader, CardContent } from '@kapwa/card';\nimport { ExternalLink } from 'lucide-react';\n\ntype ListSectionItem = {\n  id: number;\n  title: string;\n  category: string;\n};\n\ntype ListsSectionProps = {\n  title: string;\n  headerTitle: string;\n  headerSubtitle?: string;\n  externalLinkText?: string;\n  listItems: ListSectionItem[];\n};\n\nconst List = ({\n  headerTitle,\n  headerSubtitle,\n  externalLinkText,\n  listItems,\n  title,\n}: {\n  title: string;\n  headerTitle: string;\n  headerSubtitle?: string;\n  externalLinkText?: string;\n  listItems: ListSectionItem[];\n}) => {\n  return (\n    <Card>\n      <CardHeader>\n        <h2 className='text-2xl font-semibold'>{headerTitle}</h2>\n        {headerSubtitle && <p className='text-gray-800'>{headerSubtitle}</p>}\n      </CardHeader>\n      <CardContent>\n        <div className='space-y-8'>\n          <div>\n            <h3 className='font-medium mb-4'>{title}</h3>\n            <div className='space-y-4'>\n              {listItems.map(item => (\n                <div\n                  key={item.id}\n                  className='p-4 bg-white border border-gray-200 rounded-lg hover:border-primary-500 transition-colors'\n                >\n                  <div className='flex items-start justify-between'>\n                    <div>\n                      <h4 className='text-lg font-medium text-gray-900'>\n                        {item.title}\n                      </h4>\n                      <span className='inline-block px-2 py-1 mt-2 text-xs font-medium rounded-sm bg-gray-100 text-gray-800'>\n                        {item.category}\n                      </span>\n                    </div>\n                    {externalLinkText && (\n                      <Button variant='ghost' size='sm'>\n                        {externalLinkText}\n                        <ExternalLink className='ml-2 h-4 w-4' />\n                      </Button>\n                    )}\n                  </div>\n                </div>\n              ))}\n            </div>\n          </div>\n        </div>\n      </CardContent>\n    </Card>\n  );\n};\n\nexport { List };\nexport type { ListsSectionProps, ListSectionItem };\n",
       "type": "registry:component"
     }
   ],
