+++ registry.json
@@ -76,7 +76,7 @@
       "description": "A section showcasing table styles for data presentation.",
       "files": [
         {
-          "path": "src/lib/demo/blocks/tables/index.tsx",
+          "path": "src/lib/demo/blocks/sections/tables/index.tsx",
           "type": "registry:block"
         }
       ]
@@ -89,7 +89,7 @@
       "description": "A section showcasing article cards and content styles.",
       "files": [
         {
-          "path": "src/lib/demo/blocks/articles/index.tsx",
+          "path": "src/lib/demo/blocks/sections/articles/index.tsx",
           "type": "registry:block"
         }
       ]
@@ -102,7 +102,7 @@
       "description": "A section showcasing banner and call-to-action components for community engagement.",
       "files": [
         {
-          "path": "src/lib/demo/blocks/banners/index.tsx",
+          "path": "src/lib/demo/blocks/sections/banners/index.tsx",
           "type": "registry:block"
         }
       ]
@@ -112,11 +112,24 @@
       "type": "registry:block",
       "title": "Biography Section",
       "registryDependencies": ["card", "button"],
-      "description": "A section showcasing government official profile and biography layout.",
+      "description": "A section showcasing government official profile and official biography layout.",
       "files": [
         {
-          "path": "src/lib/demo/blocks/biography/index.tsx",
+          "path": "src/lib/demo/blocks/sections/biography/index.tsx",
           "type": "registry:block"
+        }
+      ]
+    },
+    {
+      "name": "official-biography",
+      "type": "registry:component",
+      "title": "Official Biography",
+      "registryDependencies": ["card", "button"],
+      "description": "A component showcasing government official profile and official biography layout.",
+      "files": [
+        {
+          "path": "src/lib/kapwa/official-biography/index.tsx",
+          "type": "registry:component"
         }
       ]
     },
@@ -128,7 +141,7 @@
       "description": "A section showcasing different list styles and search results.",
       "files": [
         {
-          "path": "src/lib/demo/blocks/lists/index.tsx",
+          "path": "src/lib/demo/blocks/sections/lists/index.tsx",
           "type": "registry:block"
         }
       ]
