+++ src/lib/demo/blocks/sections/articles/index.tsx
@@ -27,7 +27,7 @@ const ArticlesSection: React.FC<ArticlesSectionProps> = ({
   articleContent,
 }: ArticlesSectionProps) => {
   return (
-    <Card className='mb-8'>
+    <Card className='mb-2'>
       <CardHeader>
         <h2 className='text-2xl font-semibold'>{headerTitle}</h2>
         {headerSubtitle && <p className='text-gray-800'>{headerSubtitle}</p>}
