+++ src/lib/demo/pages/blocks/index.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
-import { ListsSection, ListSectionItem } from '../../blocks/lists';
-import { TablesSection, TableDataItem } from '../../blocks/tables';
+import { ListsSection, ListSectionItem } from '../../blocks/sections/lists';
+import { TablesSection, TableDataItem } from '../../blocks/sections/tables';
 import {
   ArticlesSection,
   ArticleCardItem,
   ArticleContentSection,
-} from '../../blocks/articles';
+} from '../../blocks/sections/articles';
 import {
   BannersSection,
   StripBannerItem,
   FullBannerItem,
-} from '../../blocks/banners';
-import { BiographySection, BiographyProfile } from '../../blocks/biography';
+} from '../../blocks/sections/banners';
+import { BiographySection } from '../../blocks/sections/biography';
 import { CodeBlock } from '@ui/CodeBlock';
 import { libraryInstallCommands } from '@pages/utils/install-commands';
 import { Card, CardContent } from '@kapwa/card';
@@ -115,44 +115,6 @@ const Blocks: React.FC = () => {
     },
   };
 
-  const biographyProfile: BiographyProfile = {
-    name: 'Ferdinand Marcos Jr.',
-    position: '17th President of the Republic of the Philippines',
-    servingSince: 'Serving since June 30, 2022',
-    profileImage:
-      'https://images.pexels.com/photos/2379005/pexels-photo-2379005.jpeg',
-    bannerImage:
-      'https://images.pexels.com/photos/1714455/pexels-photo-1714455.jpeg',
-    contactInfo: {
-      office: 'Malacañang Palace, Manila',
-      phone: '+63 (2) 8736 8645',
-      email: 'op@president.gov.ph',
-      website: 'www.president.gov.ph',
-    },
-    education: [
-      {
-        degree: 'Bachelor of Arts in Political Science',
-        institution: 'University of Oxford',
-        year: '1975',
-      },
-      {
-        degree: 'Master in Business Administration',
-        institution: 'Wharton School of Business',
-        year: '1979',
-      },
-    ],
-    biography: [
-      'Ferdinand "Bongbong" Romualdez Marcos Jr. serves as the 17th President of the Philippines, assuming office on June 30, 2022. As the country\'s chief executive, he leads the implementation of laws and policies aimed at national development and public welfare.',
-      'Prior to his presidency, he served in various government positions including as a Senator of the Philippines from 2010 to 2016, and as Governor of Ilocos Norte. His administration focuses on economic recovery, infrastructure development, and digital transformation of government services.',
-    ],
-    achievements: [
-      'Led the implementation of digital transformation initiatives across government agencies',
-      'Established international partnerships for economic cooperation',
-      'Launched nationwide infrastructure development programs',
-      'Reformed tax collection systems for improved efficiency',
-    ],
-  };
-
   return (
     <div className='min-h-screen bg-gray-50 py-12'>
       <div className='container mx-auto px-4'>
@@ -168,8 +130,8 @@ const Blocks: React.FC = () => {
         <Card className='mb-8'>
           <CardContent>
             <div className='grid grid-cols-1 mb-8'>
-              <p className='text-sm font-semibold'>Installation via NPM:</p>
-              <CodeBlock>{libraryInstallCommands.list.npm}</CodeBlock>
+              <p className='text-sm font-semibold'>Importing the component:</p>
+              <CodeBlock>{libraryInstallCommands.list.import}</CodeBlock>
             </div>
           </CardContent>
         </Card>
@@ -183,8 +145,8 @@ const Blocks: React.FC = () => {
         <Card className='mb-8'>
           <CardContent>
             <div className='grid grid-cols-1 mb-8'>
-              <p className='text-sm font-semibold'>Installation via NPM:</p>
-              <CodeBlock>{libraryInstallCommands.table.npm}</CodeBlock>
+              <p className='text-sm font-semibold'>Importing the component:</p>
+              <CodeBlock>{libraryInstallCommands.table.import}</CodeBlock>
             </div>
           </CardContent>
         </Card>
@@ -202,9 +164,9 @@ const Blocks: React.FC = () => {
         <Card className='mb-8'>
           <CardContent>
             <div className='grid grid-cols-1 mb-8'>
-              <p className='text-sm font-semibold'>Installation via NPM:</p>
+              <p className='text-sm font-semibold'>Importing the components:</p>
               <div className='space-y-4'>
-                <CodeBlock>{libraryInstallCommands.article.npm}</CodeBlock>
+                <CodeBlock>{libraryInstallCommands.article.import}</CodeBlock>
               </div>
             </div>
           </CardContent>
@@ -222,11 +184,13 @@ const Blocks: React.FC = () => {
         <Card className='mb-8'>
           <CardContent>
             <div className='grid grid-cols-1 mb-8'>
-              <p className='text-sm font-semibold'>Installation via NPM:</p>
+              <p className='text-sm font-semibold'>Importing the components:</p>
               <div className='space-y-4'>
-                <CodeBlock>{libraryInstallCommands.stripBanner.npm}</CodeBlock>
                 <CodeBlock>
-                  {libraryInstallCommands.civicTechBanner.npm}
+                  {libraryInstallCommands.stripBanner.import}
+                </CodeBlock>
+                <CodeBlock>
+                  {libraryInstallCommands.civicTechBanner.import}
                 </CodeBlock>
               </div>
             </div>
@@ -234,16 +198,15 @@ const Blocks: React.FC = () => {
         </Card>
         <BiographySection
           headerTitle='Official Biography'
-          headerSubtitle='Government official profile and biography layout'
-          profile={biographyProfile}
-          contactButtonText='Contact Office'
-          websiteButtonText='Visit Website'
+          headerSubtitle='Government official profile and official biography layout'
         />
         <Card className='mb-8'>
           <CardContent>
             <div className='grid grid-cols-1 mb-8'>
               <p className='text-sm font-semibold'>Installation via NPM:</p>
-              <CodeBlock>{libraryInstallCommands.biography.npm}</CodeBlock>
+              <CodeBlock>
+                {libraryInstallCommands.officialBiography.import}
+              </CodeBlock>
             </div>
           </CardContent>
         </Card>
